v2 (#10)
* Incorporate main lander page (#9)

* Rename repo, point to new git URLs

* Update deps

* Fix footer

* Create a /blog route, move lander content → primary index

* Drop styled and material-ui

* Minor style improvements

* content cleanup

* move Footer outside of content

* Lander improvements
