+++ CNAME
@@ -1,2 +1,2 @@
-kylieis.online/blog
-www.kylieis.online/blog
+kylieis.online
+www.kylieis.online
