--- content/__about/general-about.md
@@ -1,22 +1,0 @@
----
-title: 'start here'
-date: 2020-06-01 12:00:00
----
-
-<div align="center">
-
-# About Kylie Stewart
-
-Kylie Stewart is an engineer and occasional speaker. She is passionate about building tools and high-performing applications, as well as hiking with her dog, Otis. She lives in downtown Denver, Colorado.
-
-</div>
-
-## Extracurriculars
-
-- What am I doing [now](https://kylieis.online/now)?
-
-## Career
-
-Prior to development, Kylie worked as a Store Coordinator at Tesla Motors in the Mid-Atlantic, as a Volunteer Coordinator in a position with the AmeriCorps VISTA program, and as a bartender on and off in between.
-
-## Hobbies
