--- content/data/talks.js
@@ -1,96 +1,0 @@
-const talks = [
-  {
-    eventDate: '5/26/2020',
-    title: 'Demystifying urql With urql-devtools',
-    description:
-      'Using urql-devtools to better understand urql as a GraphQL client.',
-    eventName: 'GraphQL Denver',
-    eventType: 'meetup',
-    exportedSlidesUrl: '',
-    hostedSlidesUrl: '',
-    recordedPresentationUrl: '',
-  },
-  {
-    eventDate: '6/28/2018',
-    title: 'Machine Learning on the Command Line',
-    description:
-      'Using Node.js to create scripts that utilize popular machine learning algorithms.',
-    eventName: 'js.la',
-    eventType: 'meetup',
-    exportedSlidesUrl:
-      'https://github.com/kale-stew/ml-on-the-cl/blob/master/final.pdf',
-    hostedSlidesUrl: 'https://kale-stew.github.io/ml-on-the-cl',
-    recordedPresentationUrl: 'https://youtu.be/MzrDy4s8MF8',
-  },
-  {
-    eventDate: '10/25/2018',
-    title: 'How TypeScript Made Me a Better JavaScript Developer',
-    description:
-      'An overview of helpful patterns for writing TypeScript, and how they apply to a similarly-structured JavaScript code base.',
-    eventName: 'Formidable Denver Open House',
-    eventType: 'meetup',
-    exportedSlidesUrl:
-      'https://github.com/kale-stew/typescript-is-awesome/blob/master/final.pdf',
-    hostedSlidesUrl: 'https://kale-stew.github.io/typescript-is-awesome/',
-    recordedPresentationUrl: null,
-  },
-  {
-    eventDate: '11/15/2019',
-    title: 'React to React Native: How Hard Could It Be?',
-    description:
-      "Comparing the browser framework to it's native companion, and contrasting the development approaches.",
-    eventName: 'RVA.js',
-    eventType: 'conference',
-    exportedSlidesUrl: '',
-    hostedSlidesUrl: '',
-    recordedPresentationUrl: 'https://www.youtube.com/watch?v=i3DYPaHXLeo',
-  },
-  // {
-  //   "eventDate": "5/28/2019",
-  //   "title": "React to React Native: How Hard Could It Be?",
-  //   "description": "Comparing the browser framework to it's native companion, and contrasting the development approaches",
-  //   "eventName": "DenverScript",
-  //   "eventType": "meetup",
-  //   "exportedSlidesUrl": "https://github.com/kale-stew/react-to-react-native/blob/master/final.pdf",
-  //   "hostedSlidesUrl": "https://kale-stew.github.io/react-to-react-native/",
-  //   "recordedPresentationUrl": "https://youtu.be/87trwew9Hik?t=1587"
-  // },
-  {
-    eventDate: '9/20/2019',
-    title: 'Evolution of an API: A Case for GraphQL',
-    description:
-      'A discussion of API development over the years, and how GraphQL has evolved to address the problems that have arisen in that time.',
-    eventName: 'UtahJS',
-    eventType: 'conference',
-    exportedSlidesUrl:
-      'https://github.com/kale-stew/talks/blob/master/evolution-of-an-api/final.pdf',
-    hostedSlidesUrl: '',
-    recordedPresentationUrl: 'https://www.youtube.com/watch?v=RhQx579OTVc',
-  },
-  {
-    eventDate: '4/16/2019',
-    title: 'TypeScript in React: A Love Story',
-    description:
-      'An overview of practices my team found beneficial while integrating TypeScript into a React monorepo.',
-    eventName: 'React Denver',
-    eventType: 'meetup',
-    exportedSlidesUrl:
-      'https://github.com/kale-stew/ts-in-react/blob/master/final.pdf',
-    hostedSlidesUrl: 'https://kale-stew.github.io/ts-in-react/',
-    recordedPresentationUrl: 'https://youtu.be/iBlGIS-UQsw?t=1638',
-  },
-  {
-    eventDate: '4/28/2018',
-    title: 'Using JavaScript to Teach Machines How to do Cool Things',
-    description:
-      "A gentle introduction to machine learning - defining basic algorithms and discussing why we wouldn't usually use JS to implement them.",
-    eventName: 'Zeit Day',
-    eventType: 'conference',
-    exportedSlidesUrl:
-      'https://github.com/kale-stew/zeit-day-2018/blob/master/final.pdf',
-    hostedSlidesUrl: 'https://kale-stew.github.io/zeit-day-2018/',
-    recordedPresentationUrl: 'https://youtu.be/QaV7a64mUYE',
-  },
-]
-
-export default talks
