+++ package.json
@@ -1,11 +1,10 @@
 {
-  "name": "kale-stew-blog",
-  "private": true,
+  "name": "kylieis.online",
   "description": "Kylie Stewart's online blog",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "author": "Kylie Stewart <kylie@kyliestewart.tech>",
   "bugs": {
-    "url": "https://github.com/kale-stew/blog/issues"
+    "url": "https://github.com/kale-stew/kale-stew.github.io/issues"
   },
   "scripts": {
     "build": "gatsby clean && gatsby build --prefix-paths",
@@ -45,19 +44,19 @@
     "gatsby-transformer-sharp": "^2.1.17",
     "intersection-observer": "^0.5.1",
     "katex": "^0.11.1",
-    "lodash": "^4.17.11",
+    "lodash.partial": "^4.2.1",
+    "lodash.uniq": "^4.5.0",
     "node-sass": "^4.11.0",
     "prismjs": "^1.15.0",
-    "react": "^16.8.1",
-    "react-dom": "^16.8.1",
+    "react": "^16.10.2",
+    "react-animations": "^1.0.0",
+    "react-dom": "^16.10.2",
     "react-helmet": "^5.2.0",
+    "react-scripts": "3.2.0",
+    "react-social-icons": "^4.1.0",
     "react-switch": "^5.0.0",
-    "react-typography": "^0.16.13",
     "smooth-scroll": "^16.0.0",
     "typeface-fira-sans": "^0.0.75",
-    "typeface-merriweather": "0.0.72",
-    "typeface-montserrat": "0.0.54",
-    "typeface-noto-sans-kr": "0.0.63",
     "typography": "^0.16.17",
     "typography-theme-github": "^0.16.18"
   },
@@ -75,12 +74,13 @@
     "signale": "^1.3.0"
   },
   "keywords": [
-    "blog",
-    "gatsby"
+    "personal site",
+    "gatsby",
+    "blog"
   ],
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/kale-stew/blog.git"
+    "url": "git+https://github.com/kale-stew/kale-stew.github.io.git"
   }
 }
