+++ src/components/footer/index.scss
@@ -1,7 +1,25 @@
 @import '../../styles/variables.scss';
 
 .footer {
-  padding-top: 52px;
-  text-align: center;
-  font-size: 12px;
+  background: $theme-gradient;
+  position: fixed;
+  width: 100%;
+  height: 65px;
+  bottom: 0;
+  padding: 20px 12px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  & * {
+    margin-right: 1rem;
+  }
+}
+
+.footer-link {
+  color: rgb(233, 233, 233) !important;
+
+  &:hover {
+    text-decoration: underline;
+  }
 }
