+++ src/constants/enum.js
@@ -6,3 +6,5 @@ export const THEME = {
   LIGHT: 'light',
   DARK: 'dark',
 }
+
+export const SCROLL_Y = 'y'
