+++ src/constants/index.js
@@ -1,4 +1,12 @@
 export * from './enum'
 
+export const BLOG_TITLE = `Blog`
 export const HOME_TITLE = `Home`
-export const SCROLL_Y = 'y'
+
+export const SOCIALS = {
+  Email: 'kylie@kyliestewart.tech',
+  Homepage: 'https://kylieis.online',
+  Instagram: 'https://instagram.com/kalestews',
+  LinkedIn: 'https://www.linkedin.com/in/kylieastewart/',
+  Twitter: 'https://twitter.com/kyliestew',
+}
