+++ src/styles/dark-theme.scss
@@ -25,6 +25,52 @@ body.dark {
     }
   }
 
+  .lander {
+    max-width: 48rem;
+    margin: 0px auto;
+    font-family: 'Fira Sans';
+
+    h2 {
+      font-style: italic;
+      font-size: 26px;
+      align-self: center;
+    }
+  }
+
+  .intro {
+    display: flex;
+    font-size: 20px;
+    flex-direction: column;
+    justify-content: center;
+    align-content: center;
+    margin-bottom: 3rem;
+
+    @media (max-width: 1024px) {
+      margin-bottom: 0.1rem;
+    }
+
+    & * {
+      align-self: center;
+
+      &:last-child {
+        margin-bottom: 3rem;
+      }
+    }
+  }
+
+  .project-wrapper {
+    display: flex;
+    justify-content: space-evenly;
+
+    @media (max-width: 1024px) {
+      flex-direction: column;
+      align-items: center;
+      :last-child {
+        margin-bottom: 5rem;
+      }
+    }
+  }
+
   .category-container {
     border-top: 1px solid $dark-category-border-color;
     border-bottom: 1px solid $dark-category-border-color;
@@ -60,15 +106,9 @@ body.dark {
       color: $dark-middle-font-color;
     }
     h3:hover {
-      font-style: italic;
       text-decoration: underline;
       color: $dark-link-color;
     }
-  }
-
-  .footer {
-    display: flex;
-    justify-content: space-evenly;
   }
 
   // blog-post
