+++ src/styles/light-theme.scss
@@ -6,6 +6,51 @@ body.light {
     color: $light-colored-font-color;
   }
 
+  .lander {
+    max-width: 48rem;
+    margin: 0px auto;
+    font-family: 'Fira Sans';
+
+    h2 {
+      font-style: italic;
+      font-size: 26px;
+      align-self: center;
+    }
+  }
+
+  .intro {
+    display: flex;
+    font-size: 20px;
+    flex-direction: column;
+    justify-content: center;
+    align-content: center;
+    margin-bottom: 3rem;
+
+    @media (max-width: 1024px) {
+      margin-bottom: 0.1rem;
+    }
+
+    & * {
+      align-self: center;
+      &:last-child {
+        margin-bottom: 3rem;
+      }
+    }
+  }
+
+  .project-wrapper {
+    display: flex;
+    justify-content: space-evenly;
+
+    @media (max-width: 1024px) {
+      flex-direction: column;
+      align-items: center;
+      :last-child {
+        margin-bottom: 5rem;
+      }
+    }
+  }
+
   .home-header {
     color: $dark-gray;
   }
@@ -47,18 +92,12 @@ body.light {
       color: $dark-gray;
     }
     h3:hover {
-      font-style: italic;
       text-decoration: underline;
       color: $light-link-color;
     }
     p {
       color: $middle-gray;
     }
-  }
-
-  .footer {
-    display: flex;
-    justify-content: space-evenly;
   }
 
   // blog-post
