+++ src/utils/storage/localStorage.js
@@ -1,7 +1,6 @@
-import _ from 'lodash'
-
+import partial from 'lodash.partial'
 import { setValueTo, getValueFrom } from './core'
 import { localStorage } from './browser'
 
-export const setValueToLocalStorage = _.partial(setValueTo, localStorage)
-export const getValueFromLocalStorage = _.partial(getValueFrom, localStorage)
+export const setValueToLocalStorage = partial(setValueTo, localStorage)
+export const getValueFromLocalStorage = partial(getValueFrom, localStorage)
